Create How many are even.java
Write a function that accepts 5 input parameters and returns the count of how many those 5 
are even.
For example,
If the five input parameters are 12, 17, 19, 14, and 115, there are two even numbers 12  and 
14. So, the function must return 2.

Similarly,
If the five input parameters are 15, 0, -12, 19, and 28, there are three even numbers 0, -12, 
and 28. So, the function must return 3.Observe that zero is also considered an even number.